Add files via upload
The increasing popularity of online shopping has led to the emergence of new economic activities. To succeed in the highly competitive e-commerce environment, it is vital to understand consumer intention. Understanding what motivates consumer intention is critical because such intention is key to survival in this fast-paced and hyper competitive environment. In this kernel, I have predicted whether a visitor to a web page has the potential to create a sales conversion or not. I have utilized extensive feature engineering techniques and created 90% accurate models via hyperparameter tuning